Remove VMTInclude + Separate DHCP paquets by class
