+++ source/Cosmos.System2/Network/IPV4/UDP/DHCP/DHCPPacket.cs
@@ -61,14 +61,6 @@ namespace Cosmos.System.Network.IPv4.UDP.DHCP
         }
 
         /// <summary>
-        /// Work around to make VMT scanner include the initFields method
-        /// </summary>
-        public static void VMTInclude()
-        {
-            new DHCPPacket();
-        }
-
-        /// <summary>
         /// Create new inctanse of the <see cref="DHCPPacket"/> class.
         /// </summary>
         internal DHCPPacket() : base()
@@ -215,282 +207,6 @@ namespace Cosmos.System.Network.IPv4.UDP.DHCP
         /// Get DHCP Options
         /// </summary>
         internal List<DHCPOption> Options { get; private set; }
-
-    }
-
-    /// <summary>
-    /// DHCPDiscover class.
-    /// </summary>
-    internal class DHCPDiscover : DHCPPacket
-    {
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPDiscover"/> class.
-        /// </summary>
-        internal DHCPDiscover() : base()
-        { }
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPDiscover"/> class.
-        /// </summary>
-        /// <param name="rawData">Raw data.</param>
-        internal DHCPDiscover(byte[] rawData) : base(rawData)
-        { }
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPDiscover"/> class.
-        /// </summary>
-        /// <param name="mac_src">Source MAC Address.</param>
-        /// <exception cref="ArgumentException">Thrown if RawData is invalid or null.</exception>
-        internal DHCPDiscover(MACAddress mac_src) : base(mac_src, 10) //discover packet size
-        {
-            //Discover
-            RawData[282] = 0x35;
-            RawData[283] = 0x01;
-            RawData[284] = 0x01;
-
-            //Parameters start here
-            RawData[285] = 0x37;
-            RawData[286] = 4;
-
-            //Parameters*
-            RawData[287] = 0x01;
-            RawData[288] = 0x03;
-            RawData[289] = 0x0f;
-            RawData[290] = 0x06;
-
-            RawData[291] = 0xff; //ENDMARK
-        }
-
-        /// <summary>
-        /// Work around to make VMT scanner include the initFields method
-        /// </summary>
-        public new static void VMTInclude()
-        {
-            new DHCPDiscover();
-        }
-
-        protected override void initFields()
-        {
-            base.initFields();
-        }
-
-    }
-
-    /// <summary>
-    /// DHCPRequest class.
-    /// </summary>
-    internal class DHCPRequest : DHCPPacket
-    {
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPRequest"/> class.
-        /// </summary>
-        internal DHCPRequest() : base()
-        { }
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPRequest"/> class.
-        /// </summary>
-        /// <param name="rawData">Raw data.</param>
-        internal DHCPRequest(byte[] rawData) : base(rawData)
-        { }
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPRequest"/> class.
-        /// </summary>
-        /// <param name="mac_src">Source MAC Address.</param>
-        /// <param name="RequestedAddress">Requested Address.</param>
-        /// <param name="DHCPServerAddress">DHCP server IPv4 Address.</param>
-        /// <exception cref="OverflowException">Thrown if data array length is greater than Int32.MaxValue.</exception>
-        /// <exception cref="ArgumentException">Thrown if RawData is invalid or null.</exception>
-        internal DHCPRequest(MACAddress mac_src, Address RequestedAddress, Address DHCPServerAddress) : base(mac_src, 22)
-        {
-            //Request
-            RawData[282] = 53;
-            RawData[283] = 1;
-            RawData[284] = 3;
-
-            //Requested Address
-            RawData[285] = 50;
-            RawData[286] = 4;
-
-            RawData[287] = RequestedAddress.address[0];
-            RawData[288] = RequestedAddress.address[1];
-            RawData[289] = RequestedAddress.address[2];
-            RawData[290] = RequestedAddress.address[3];
-
-            RawData[291] = 54;
-            RawData[292] = 4;
-
-            RawData[293] = DHCPServerAddress.address[0];
-            RawData[294] = DHCPServerAddress.address[1];
-            RawData[295] = DHCPServerAddress.address[2];
-            RawData[296] = DHCPServerAddress.address[3];
-
-            //Parameters start here
-            RawData[297] = 0x37;
-            RawData[298] = 4;
-
-            //Parameters
-            RawData[299] = 0x01;
-            RawData[300] = 0x03;
-            RawData[301] = 0x0f;
-            RawData[302] = 0x06;
-
-            RawData[303] = 0xff; //ENDMARK
-        }
-
-        /// <summary>
-        /// Work around to make VMT scanner include the initFields method
-        /// </summary>
-        public new static void VMTInclude()
-        {
-            new DHCPRequest();
-        }
-
-        /// <summary>
-        /// Init DHCPRequest fields.
-        /// </summary>
-        /// <exception cref="ArgumentException">Thrown if RawData is invalid or null.</exception>
-        protected override void initFields()
-        {
-            base.initFields();
-        }
-
-    }
-
-    /// <summary>
-    /// DHCPAck class.
-    /// </summary>
-    internal class DHCPAck : DHCPPacket
-    {
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPAck"/> class.
-        /// </summary>
-        internal DHCPAck() : base()
-        { }
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPAck"/> class.
-        /// </summary>
-        /// <param name="rawData">Raw data.</param>
-        internal DHCPAck(byte[] rawData) : base(rawData)
-        { }
-
-        /// <summary>
-        /// Work around to make VMT scanner include the initFields method
-        /// </summary>
-        public new static void VMTInclude()
-        {
-            new DHCPAck();
-        }
-
-        /// <summary>
-        /// Init DHCPAck fields.
-        /// </summary>
-        /// <exception cref="ArgumentException">Thrown if RawData is invalid or null.</exception>
-        protected override void initFields()
-        {
-            base.initFields();
-
-            foreach (var option in Options)
-            {
-                if (option.Type == 1) //Mask
-                {
-                    Subnet = new Address(option.Data, 0);
-                }
-                else if (option.Type == 6) //DNS
-                {
-                    DNS = new Address(option.Data, 0);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Get Subnet IPv4 Address
-        /// </summary>
-        internal Address Subnet { get; private set; }
-
-        /// <summary>
-        /// Get DNS IPv4 Address
-        /// </summary>
-        internal Address DNS { get; private set; }
-
-    }
-
-    /// <summary>
-    /// DHCPRelease class.
-    /// </summary>
-    internal class DHCPRelease : DHCPPacket
-    {
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPRelease"/> class.
-        /// </summary>
-        internal DHCPRelease() : base()
-        { }
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPRelease"/> class.
-        /// </summary>
-        /// <param name="rawData">Raw data.</param>
-        internal DHCPRelease(byte[] rawData) : base(rawData)
-        { }
-
-        /// <summary>
-        /// Create new inctanse of the <see cref="DHCPRelease"/> class.
-        /// </summary>
-        /// <param name="client">Client IPv4 Address.</param>
-        /// <param name="server">DHCP Server IPv4 Address.</param>
-        /// <param name="source">Source MAC Address.</param>
-        /// <exception cref="OverflowException">Thrown if data array length is greater than Int32.MaxValue.</exception>
-        /// <exception cref="ArgumentException">Thrown if RawData is invalid or null.</exception>
-        internal DHCPRelease(Address client, Address server, MACAddress source) : base(client, server, source, 19)
-        {
-            //Release
-            RawData[282] = 0x35;
-            RawData[283] = 0x01;
-            RawData[284] = 0x07;
-
-            //DHCP Server ID
-            RawData[285] = 0x36;
-            RawData[286] = 0x04;
-
-            RawData[287] = server.address[0];
-            RawData[288] = server.address[1];
-            RawData[289] = server.address[2];
-            RawData[290] = server.address[3];
-
-            //Client ID
-            RawData[291] = 0x3d;
-            RawData[292] = 7;
-            RawData[293] = 1;
-
-            RawData[294] = source.bytes[0];
-            RawData[295] = source.bytes[1];
-            RawData[296] = source.bytes[2];
-            RawData[297] = source.bytes[3];
-            RawData[298] = source.bytes[4];
-            RawData[299] = source.bytes[5];
-
-            RawData[300] = 0xff; //ENDMARK
-        }
-
-        /// <summary>
-        /// Work around to make VMT scanner include the initFields method
-        /// </summary>
-        public new static void VMTInclude()
-        {
-            new DHCPRelease();
-        }
-
-        /// <summary>
-        /// Init DHCPRelease fields.
-        /// </summary>
-        /// <exception cref="ArgumentException">Thrown if RawData is invalid or null.</exception>
-        protected override void initFields()
-        {
-            base.initFields();
-        }
 
     }
 }
