+++ source/Cosmos.System2/Network/IPv4/UDP/UDPPacket.cs
@@ -54,82 +54,6 @@ namespace Cosmos.System.Network.IPv4.UDP
         }
 
         /// <summary>
-        /// Make header.
-        /// </summary>
-        /// <param name="sourceIP">Source IP.</param>
-        /// <param name="destIP">Destination IP.</param>
-        /// <param name="udpLen">UDP length.</param>
-        /// <param name="sourcePort">Source port.</param>
-        /// <param name="destPort">Destination port.</param>
-        /// <param name="UDP_Data">UDP data.</param>
-        /// <returns>byte array value.</returns>
-        /// <exception cref="OverflowException">Thrown if UDP_Data array length is greater than Int32.MaxValue.</exception>
-        public static byte[] MakeHeader(byte[] sourceIP, byte[] destIP, UInt16 udpLen, UInt16 sourcePort, UInt16 destPort, byte[] UDP_Data)
-        {
-            byte[] header = new byte[18 + UDP_Data.Length];
-
-            header[0] = sourceIP[0];
-            header[1] = sourceIP[1];
-            header[2] = sourceIP[2];
-            header[3] = sourceIP[3];
-
-            header[4] = destIP[0];
-            header[5] = destIP[1];
-            header[6] = destIP[2];
-            header[7] = destIP[3];
-
-            header[8] = 0x00;
-
-            header[9] = 0x11;
-
-            header[10] = (byte)((udpLen >> 8) & 0xFF);
-            header[11] = (byte)((udpLen >> 0) & 0xFF);
-
-            header[12] = (byte)((sourcePort >> 8) & 0xFF);
-            header[13] = (byte)((sourcePort >> 0) & 0xFF);
-
-            header[14] = (byte)((destPort >> 8) & 0xFF);
-            header[15] = (byte)((destPort >> 0) & 0xFF);
-
-            header[16] = (byte)((udpLen >> 8) & 0xFF);
-            header[17] = (byte)((udpLen >> 0) & 0xFF);
-
-            for (int i = 0; i < UDP_Data.Length; i++)
-            {
-                header[18 + i] = UDP_Data[i];
-            }
-
-            return header;
-        }
-
-        /// <summary>
-        /// Check CRC.
-        /// </summary>
-        /// <param name="packet">Packer to be checked.</param>
-        /// <returns>bool value.</returns>
-        /// <exception cref="OverflowException">Thrown if packet.UDP_Data lenght if greater than Int32.MaxValue.</exception>
-        public static bool CheckCRC(UDPPacket packet)
-        {
-            byte[] header = MakeHeader(packet.SourceIP.address, packet.DestinationIP.address, packet.UDP_Length, packet.SourcePort, packet.DestinationPort, packet.UDP_Data);
-            if (CalcOcCRC(header, 0, header.Length) == packet.udpCRC)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        /// <summary>
-        /// Work around to make VMT scanner include the initFields method
-        /// </summary>
-        public static void VMTInclude()
-        {
-            new UDPPacket();
-        }
-
-        /// <summary>
         /// Create new inctanse of the <see cref="UDPPacket"/> class.
         /// </summary>
         internal UDPPacket()
